make controllers and routes connected with token

+++ controllers/contact.js
@@ -2,8 +2,12 @@ const Contact = require('../models/contact')
 
 const createContact_post = async (req, res) => {
   try {
-    const contact = await Contact.create(req.body)
-    res.send(contact)
+
+    if (res.locals.payload.id === req.body.userId) {
+      const contact = await Contact.create(req.body)
+      res.status(201).send({ status: "issue Created successfully", contact })
+    }
+    res.status(400).send("faild to create issue")
   } catch (error) {
     res.status(401).send({
       status: 'Error',
@@ -15,8 +19,12 @@ const createContact_post = async (req, res) => {
 
 const get_all_Contacts_get = async (req, res) => {
   try {
-    const contacts = await Contact.find()
-    res.send(contacts)
+
+    if (res.locals.payload.role === "admin") {
+      const contacts = await Contact.find()
+      res.send(contacts)
+    }
+    res.status(400).send("faild to Update issue")
   } catch (error) {
     res.status(401).send({
       status: 'Error',
@@ -27,8 +35,11 @@ const get_all_Contacts_get = async (req, res) => {
 }
 const get_Contact_put = async (req, res) => {
   try {
-    const contact = await Contact.findByIdAndUpdate(req.params.contactId)
-    res.send(contact)
+    if (res.locals.payload.role === "admin") {
+      const contact = await Contact.findByIdAndUpdate(req.params.contactId)
+      res.status(200).send({ status: "issue Update Successfully", contact })
+    }
+    res.status(400)
   } catch (error) {
     res.status(401).send({
       status: 'Error',
@@ -39,8 +50,11 @@ const get_Contact_put = async (req, res) => {
 }
 const get_Contact_delete = async (req, res) => {
   try {
-    const contact = await Contact.findByIdAndDelete(req.params.contactId)
-    res.send(contact)
+
+    if (res.locals.payload.role === "admin") {
+      const contact = await Contact.findByIdAndDelete(req.params.contactId)
+      res.status(200).send({ status: "issue delete!", contact })
+    }
   } catch (error) {
     res.status(401).send({
       status: 'Error',
